requirements.txt
PyNaCl
requests
websocket-client
numpy
python-dotenv
rich
websockets
sympy==1.12